pyproject.toml
[project]
name = "Smart-Support-System"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Saravanan", email = "saravanansd634@gmail.com"},
]
dependencies = ["fastapi>=0.116.1", "uvicorn>=0.35.0", "python-dotenv>=1.1.1", "pydantic>=2.11.7", "langchain>=0.3.27", "langchain-groq>=0.3.6", "celery[redis]>=5.5.3", "redis>=5.2.1", "groq>=0.30.0", "motor>=3.7.1"]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
